Salamander Packets initial source push
Making an initial push to the repo to start filling in some of the
implementation details of packet processing.

For right now we are putting all of the packet code in a single source
file, separated into a couple of different modules... For packet
handles, packet views, and protocol handling.

empty lib.rs for compilation checking.