Packets - basic test for PacketView
Basic sanity checks for some of the PacketView functionality.  Starting
small by just making sure that we can instantiate a new PacketView, that
it defaults nicely, and that we can add a data source and retrieve the
associated data.  Still trying to get a handle on some of the rust
paradigms, especially borrowing, but we now have code that does something.